try fix

+++ app/Http/Controllers/CartController.php
@@ -37,7 +37,7 @@ class CartController extends Controller
 
     public function update(Request $request, CartItem $cartItem)
     {
-        if ($cartItem->user_id !== Auth::id()) {
+        if ((int) $cartItem->user_id !== (int) Auth::id()) {
             abort(403);
         }
 
@@ -52,7 +52,7 @@ class CartController extends Controller
 
     public function destroy(CartItem $cartItem)
     {
-        if ($cartItem->user_id !== Auth::id()) {
+        if ((int) $cartItem->user_id !== (int) Auth::id()) {
             abort(403);
         }
